+++ src/zl-util/STLString.h
@@ -28,12 +28,12 @@ class STLString :
 public:
 
 	//----------------------------------------------------------------//
-	void		base_64_decode		( void* buffer, u32 len );
-	void		base_64_encode		( const void* buffer, u32 len );
-	STLString	clip				( u32 first, u32 last );
-	STLString	clip_to_back		( u32 first );
-	STLString	clip_to_front		( u32 last );
-	void		hex_encode			( const void* buffer, u32 len );
+	void		base_64_decode		( void* buffer, size_t len );
+	void		base_64_encode		( const void* buffer, size_t len );
+	STLString	clip				( size_t first, size_t last );
+	STLString	clip_to_back		( size_t first );
+	STLString	clip_to_front		( size_t last );
+	void		hex_encode			( const void* buffer, size_t len );
 	static u8	hex_to_byte			( u32 c );
 	void		replace_char		( cc8 match, cc8 sub );
 	void		tokenize			( STLArray < STLString >& tokens, const STLString& delimiters = " " );
